解决至尊娱乐后台玩家列表>设置玩家身份，在身份类型的赢玩家和输玩家中，添加玩家赢概率和输概率功能

+++ Application/Admin/Controller/UserController.class.php
@@ -1667,7 +1667,7 @@ class UserController extends AdminController
             $type = I('type');
             $statusValue = I('statusValue');
             $moneyLimit  = I('moneyLimit');
-            $ctrlPercent  = (int)I('ctrlPercent');
+            $ctrlPercent  = I('ctrlPercent');
             if ($statusValue == 0) {
                 $this->error('请选择需要设置的身份');
             }
